add moderation page, votes for words feature, redis pre-config

+++ backend/src/main/java/com/example/dictionary/repositories/WordRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WordRepository extends JpaRepository<Word,Integer> {
+public interface WordRepository extends JpaRepository<Word, Long> {
 
     Word findByValueEquals(String value);
     Word findByTranslateEquals(String translate);
